feat: create robot ui directory
